Refactoring, Post Image, Middleware pending (issue with token authorization)
