+++ Museic-functions/.vscode/tasks.json
@@ -1,0 +1,13 @@
+{
+	"version": "2.0.0",
+	"tasks": [
+		{
+			"type": "npm",
+			"script": "deploy",
+			"path": "functions/",
+			"problemMatcher": [],
+			"label": "npm: deploy - functions",
+			"detail": "firebase deploy --only functions"
+		}
+	]
+}
