+++ Museic-functions/functions/utilidades/administrador.js
@@ -1,0 +1,7 @@
+const admin = require("firebase-admin");
+
+admin.initializeApp();
+
+const db = admin.firestore();
+
+module.exports = { admin, db };
