+++ Museic-functions/functions/utilidades/configuracion.js
@@ -1,0 +1,10 @@
+module.exports = {
+    apiKey: "AIzaSyADNStdDaRzHgMc9TtyR0-_8LBd2YR15hI",
+    authDomain: "muse-ic.firebaseapp.com",
+    databaseURL: "https://muse-ic.firebaseio.com",
+    projectId: "muse-ic",
+    storageBucket: "muse-ic.appspot.com",
+    messagingSenderId: "461054247985",
+    appId: "1:461054247985:web:14f138f643aa4d9c4c4110",
+    measurementId: "G-E1YDHYL8XW",
+  };
